requirements.txt
aiogram==3.13.1
asyncpg==0.29.0
groq==0.11.0
aiohttp==3.10.5
python-dotenv==1.0.1
